requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.10.0
anthropic>=0.10.0
openai>=1.10.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
numpy>=1.22.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
nltk>=3.8.0
pydantic>=2.0.0
tqdm>=4.64.0 